这一版改动非常大， 主要的改动是，增加了一个特殊的接口，目前当前接口只是个测试版，叫做 callstacks 。 ADD 增加功能，调试器进程创建的部分可以选择是否等待调试器进程结束。 ADD 增加了一个 EVENT 库，叫做 EventHelper.py 具体功能里面自己看 FIX 调整整体代码结构，增加了一个分发器模块 FIX 读文件的位置增加一个忽略异常的参数 ADD 综上，增加了一个功能，功能是，可以让脚本根据之前调试过程中得到的结果决定后续的调试流程，基于此，后续可以做完全的动态调试了，不仅仅是静态内容输出了 ADD 增加了一个DLL，叫做 zoo_event.dll ，功能是，这个dll给WinDBG增加了两个Event 参数，让调试器可以和脚本交互

+++ WinDBG/Lib/BaseLibrary.py
@@ -104,9 +104,9 @@ def RunProcess(exe, cmd):
 
 # 创建进程并且等待返回
 
-def RunProcessWaitReturn(exe, cmd):
+def RunProcessWaitReturn(exe, cmd, wait=True):
     handle = RunProcess(exe, cmd)
-    if handle is not None:
+    if handle is not None and wait:
         win32event.WaitForSingleObject(handle[0], -1)
 
 
@@ -150,41 +150,42 @@ def MakeDebugToolPath():
 
 # 执行一个脚本文件
 
-def RunFileWithDebuger(dump, file):
+def RunFileWithDebuger(dump, file, wait=True):
     script_path = file
     debug_tool = MakeDebugToolPath()
     debug_cmd = MakeScriptCommand(dump, script_path)
     try:
-        RunProcessWaitReturn(debug_tool, debug_cmd)
+        RunProcessWaitReturn(debug_tool, debug_cmd, wait)
         pass
     except Exception as e:
         print(e)
-    os.unlink(script_path)
+    if wait:
+        os.unlink(script_path)
     pass
 
 
 # 执行一段脚本
 
-def RunScriptWithDebuger(dump, script):
-    RunFileWithDebuger(dump, SaveStringToTempFile(script))
+def RunScriptWithDebuger(dump, script, wait=True):
+    RunFileWithDebuger(dump, SaveStringToTempFile(script), wait)
 
 
 # 执行一条命令
 
-def RunCommandWithDebuger(dump, cmd, out_path):
+def RunCommandWithDebuger(dump, cmd, out_path, wait=True):
     script_text = '''
 .logopen ''' + out_path + '''
 ''' + cmd + '''
 .logclose
 q
 '''
-    RunScriptWithDebuger(dump, script_text)
+    RunScriptWithDebuger(dump, script_text, wait)
     pass
 
 
 # 执行多条命令
 
-def RunLotCommandWithDebuger(dump, cmds, out_path):
+def RunLotCommandWithDebuger(dump, cmds, out_path, wait=True):
     cmd = ""
     for line in cmds:
         if cmd == "":
@@ -194,14 +195,14 @@ def RunLotCommandWithDebuger(dump, cmds, out_path):
                 cmd += "\n"
         cmd += line
 
-    RunCommandWithDebuger(dump, cmd, out_path)
+    RunCommandWithDebuger(dump, cmd, out_path, wait)
 
 
 # 执行一个命令文件
 
-def RunCommandFileWithDebuger(dump, infile, outfile):
+def RunCommandFileWithDebuger(dump, infile, outfile, wait=True):
     cmds = LoadFileToArray(infile)
-    RunLotCommandWithDebuger(dump, cmds, outfile)
+    RunLotCommandWithDebuger(dump, cmds, outfile, wait)
     return outfile
 
 
@@ -218,6 +219,28 @@ def RemoveFileLogInfo(file):
         # 第一行去掉
         if start == 1 and line.startswith("Opened log file "):
             continue
+        # 加载wait 插件的模块去掉
+        if start == 2 and line.startswith("kd> .load ") and line.endswith("zoo_event.dll"):
+            continue
+
+        # set 功能函数去掉
+        if line.startswith("kd> !zoo_set"):
+            continue
+        if line.startswith("Info:    Set Event Name = "):
+            continue
+        if line.startswith("Info:    Set Event "):
+            continue
+        # wait 功能函数去掉
+        if line.startswith("kd> !zoo_wait"):
+            continue
+        if line.startswith("Info:    Wait Event Name = "):
+            continue
+        if line.startswith("Info:    Wait Event "):
+            continue
+        # 加载脚本功能去掉
+        if line.startswith("kd> $$>< "):
+            continue
+
         # 倒数第三行
         if start == end - 2 and "> .logclose" in line:
             continue
